feat: AtCoder 赛时 AC 统计、站管冻结不活跃与全站排行修正
- AtCoder 比赛记录改用提交 API 统计赛时 unique AC（history/json 无 AC 字段）
- profile 增加 ForceDormant / inactiveDays 筛选，回拨 last_login_at 跳过豁免
- 过题榜全站聚合与全部时间走 user_ac_problems；缓存 key 升级
- 修复 problem 队列并发 int/int64 编译错误

+++ app/core_data/internal/biz/service/problem.go
@@ -1692,8 +1692,8 @@ func (uc *ProblemUseCase) queueStats() []struct {
 		conc int64
 		stat string
 	}{
-		{"problem_fetch", problemFetchConcurrency, model.ProblemStatusPending},
-		{"problem_analyze", problemAnalyzeConcurrency, model.ProblemStatusTagging},
+		{"problem_fetch", int64(problemFetchConcurrency), model.ProblemStatusPending},
+		{"problem_analyze", int64(problemAnalyzeConcurrency), model.ProblemStatusTagging},
 	} {
 		var msgs, consumers int64
 		inspected := false
